Move GameSupport* classes up a couple folders
