+++ servers/game-support/server/build.gradle
@@ -4,7 +4,7 @@ plugins {
 }
 
 archivesBaseName = "game-support-server"
-mainClassName = 'org.triplea.server.error.reporting.GameSupportServerApplication'
+mainClassName = 'org.triplea.server.GameSupportServerApplication'
 ext {
     releasesDir = file("$buildDir/releases")
 }
